Eliminate duplicated code
Move common code related to Django REST to utils.rest.
