+++ web/attempts/rest.py
@@ -1,29 +1,16 @@
-import json
 from django.db import transaction
 from rest_framework import validators, decorators, status
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.serializers import ModelSerializer, Field
 from rest_framework.viewsets import ModelViewSet
+from utils.rest import JSONStringField
 from .models import Attempt
 
 
 class WritableJSONField(Field):
     def to_internal_value(self, data):
         return data
-
-
-class JSONStringField(Field):
-    """
-    Store a JSON object in a TextField.
-    When object is received store its json dump.
-    When object is retrieved load JSON object from string representation.
-    """
-    def to_internal_value(self, data):
-        return json.dumps(data)
-
-    def to_representation(self, value):
-        return json.loads(value)
 
 
 class AttemptSerializer(ModelSerializer):
