+++ web/problems/rest.py
@@ -1,21 +1,7 @@
-import json
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.serializers import ModelSerializer
-from rest_framework.fields import Field
+from utils.rest import JSONStringField
 from .models import Problem, Part
-
-
-class JSONStringField(Field):
-    """
-    Store a JSON object in a TextField.
-    When object is received store its json dump.
-    When object is retrieved load JSON object from string representation.
-    """
-    def to_internal_value(self, data):
-        return json.dumps(data)
-
-    def to_representation(self, value):
-        return json.loads(value)
 
 
 class PartSerializer(ModelSerializer):
